Add --opamp-api-url flag to unix install script
This commit adds a new flag to the install script, responsible for
setting the OpAmp API URL.

Previously, the OpAmp API URL was computed from the API URL. However,
this did not work correctly in all scenarios.

This change is related to a pending change in the agent and may need to
be held back until the agent change is committed.